# build: Introduce optimized profile (20x faster incremental build, 20-50% slower runtime performance compared to Release)