Motivation is to have a build that is fast for local iteration, especially incremental builds, but still has more runtime performance than debug builds.

My test case to measure incremental build time after touching a single file:
```
$ cargo build --profile optimized --bin materialized
$ touch src/adapter/src/lib.rs
$ cargo build --profile optimized --bin materialized
    Finished `optimized` profile [optimized] target(s) in 18.89s
```
vs release:
```
$ cargo build --release --bin materialized
$ touch src/adapter/src/lib.rs
$ cargo build --release --bin materialized
    Finished `release` profile [optimized + debuginfo] target(s) in 7m 08s
```
vs dev:
```
$ cargo build --bin materialized
$ touch src/adapter/src/lib.rs
$ cargo build --bin materialized
    Finished `dev` profile [unoptimized + debuginfo] target(s) in 20.75s
```
I'm not sure if the performance is good enough. Edit:

Quick benchmark:
```
materialize=> SELECT count(*) FROM generate_series(1, 10000000);
  count
----------
 10000000
(1 row)

Time: 1106.650 ms (00:01.107)
```
vs release:
```
materialize=> SELECT count(*) FROM generate_series(1, 10000000);
  count
----------
 10000000
(1 row)

Time: 736.516 ms
```
vs dev:
```
materialize=> SELECT count(*) FROM generate_series(1, 10000000);
  count
----------
 10000000
(1 row)

Time: 6434.752 ms (00:06.435)
```

I'm not sure if we should switch the default to optimized for local building, it would make life easier for most, but if you actually need the maximum performance you'll have to manually specify `--release`. Release also has the advantage of often having builds on Dockerhub already. Taking the Release build when it's on Dockerhub, but building with Optimized otherwise would be confusing since you can get a container with variable configuration.
One suggestion would be to switch the default to `--optimized` for the `bin/{environmentd,sqllogictest}` executables, but not for `bin/mzcompose`. On the other hand, this doesn't have debug symbols, so a bit more annoying than the current default of Debug build in `run.py`. No changes necessary for now I guess.

### Checklist

- [ ] This PR has adequate test coverage / QA involvement has been duly considered. ([trigger-ci for additional test/nightly runs](https://trigger-ci.dev.materialize.com/))
- [ ] This PR has an associated up-to-date [design doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md), is a design doc ([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)), or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [ ] If this PR evolves [an existing `$T ⇔ Proto$T` mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md) (possibly in a backwards-incompatible way), then it is tagged with a `T-proto` label.
- [ ] If this PR will require changes to cloud orchestration or tests, there is a companion cloud PR to account for those changes that is tagged with the release-blocker label ([example](https://github.com/MaterializeInc/cloud/pull/5021)).
  <!-- Ask in #team-cloud on Slack if you need help preparing the cloud PR. -->
- [ ] If this PR includes major [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note), I have pinged the relevant PM to schedule a changelog post.