@def- 2024-12-05
I did a quick hacky test run in CI to see how this performs, and all tests seem to be passing: https://buildkite.com/materialize/test/builds/95832 & https://buildkite.com/materialize/nightly/builds/10620
So this should be good for fast local iteration.